Remove graspologic pip installation that no longer seems to be used
As far as I can tell this is no longer actually used and its usage was removed in this commit:
https://github.com/HKUDS/LightRAG/commit/cc36cedb070ee99092c55fd9e45f0c31b60e3b55#diff-a346bcfb05aab0cc0c0baa6018976f4efab339e8cade9f6f8fb658fcbd54ae2e

Our systems are flagging this package as having a dependency on a package with a less permissive license so I would appreciate if it can be removed if its no longer needed.  Let me know if that is not the case.